third_party/blink/web_tests/fast/dom/shadow/update-text-of-style-in-shadow-dom-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Tests update the text of styles updates styles in shadow dom, crbug.com/247280

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS borderColor is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
